# Try to speed up parse_op_key