@kgaillot 2020-10-14
I think these are decent results. There are relatively few tests remaining that have a significant increase, and most of those are in the 5-7% range for a tiny fraction of a second (presumably these are 1000-iteration runs).